メールフォームをコンポーネントとして独立させる
